requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
apispec==6.8.2
attrs==25.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flask-openapi3==4.2.1
flask-openapi3-swagger==5.27.1
flask-restx==1.3.0
flask-smorest==0.46.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
marshmallow==4.0.0
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pytz==2025.2
referencing==0.36.2
rpds-py==0.27.0
SQLAlchemy==2.0.43
SQLAlchemy-Utils==0.41.2
typing-inspection==0.4.1
typing_extensions==4.14.1
webargs==8.7.0
Werkzeug==3.1.3
